backend/requirements.txt
alembic==1.7.5
APScheduler==3.8.1
attrs==21.2.0
black==21.12b0
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.2
py==1.11.0
PyJWT==2.3.0
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.27
toml==0.10.2
tomli==1.2.2
typing-extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
Werkzeug==2.0.2
